% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergences.R
\name{add_divergence}
\alias{add_divergence}
\title{Add a statistical divergence function.}
\usage{
add_divergence(name, divergence)
}
\arguments{
\item{name}{A character string with the name of the divergence.}

\item{divergence}{A function with arguments (pre_change_mean, pre_change_sd,
post_change_mean, post_change_sd), that returns the measured divergence.}
}
\description{
This function is used to specify a divergence function beyond the three
already implemented ones.
}
\details{
After calling this function, your implemented divergence function can used in
\code{\link{tpca}} by the argument \code{divergence = name}.
}
\examples{
simple_divergence <- function(mu1, sigma1, mu2, sigma2) {
  (mu2 - mu1)^2 + sigma2^2 / sigma1^2
}
add_divergence('simple', simple_divergence)
tpca(generate_cov_mat(10), divergence = 'simple')
}
