% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpca_helpers.R
\name{pca}
\alias{pca}
\title{Principal component analysis}
\usage{
pca(cov_mat, axes = 1:data_dim)
}
\arguments{
\item{cov_mat}{A covariance matrix.}

\item{axes}{A vector indicating which principal axes are wanted.}
}
\value{
\code{pca} returns an S3 object of class "pca". This is a list with 
the following components:
\describe{
  \item{\code{vectors}}{A matrix with the chosen principal axes/eigenvectors as rows.}
  \item{\code{values}}{A vector of the corresponding eigenvalues}
}
}
\description{
Performs efficient eigendecomposition of an input covariance matrix based on
which principal axes that are wanted. If all axes are wanted, \code{\link{svd}}
is used. \code{\link{RSpectra::eigs_sym}} is used if only the 
highest or lowest eigenvalues with corresponding eigenvectors are requested.
}
