% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mat_generators.R
\name{rcor_mat}
\alias{rcor_mat}
\title{Generate a random correlation matrix}
\usage{
rcor_mat(d, k0 = d, alphad = 1)
}
\arguments{
\item{d}{An integer specifying the dimension of the correlation matrix.}

\item{k0}{An integer. d - k0 are the number of independent dimensions.}

\item{alphad}{A positive real number. See \code{\link{clusterGeneration::rcorrmatrix}}.}
}
\value{
A correlation matrix with an attribute 'which_dims_cor', indicating 
which dimensions that are the correlated ones.
}
\description{
A wrapper for \code{\link{clusterGeneration::rcorrmatrix}}, but with the possibility of
setting some dimensions to being independent of the rest.
}
