% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mat_generators.R
\name{rcor_mat_est}
\alias{rcor_mat_est}
\title{Generate a random correlation matrix estimate}
\usage{
rcor_mat_est(d, k0 = d, n = 2 * d, ...)
}
\arguments{
\item{d}{An integer specifying the dimension of the correlation matrix.}

\item{k0}{An integer. d - k0 are the number of independent dimensions.}

\item{n}{An integer specifying the number of samples the estimate is based on.}

\item{...}{Other arguments to \code{\link{rcor_mat}}.}
}
\value{
A correlation matrix estimate with an attribute 'n_obs', indicating 
how many samples the estimate is based on.
}
\description{
Generates a random correlation matrix by \eqn{R} \code{\link{rcor_mat}} 
before drawing a Wishart matrix with parameters \eqn{R} and n.
Lastly, the Wishart matrix is standardized to a correlation matrix.
}
