% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mat_generators.R
\name{rcov_mat}
\alias{rcov_mat}
\title{Generate a random covariance matrix}
\usage{
rcov_mat(d, k0 = d, range_sd = c(0.2, 5), ...)
}
\arguments{
\item{d}{An integer specifying the dimension of the correlation matrix.}

\item{k0}{An integer. d - k0 are the number of independent dimensions.}

\item{range_sd}{A vector of length 2 specifying the lower and upper bound
of the uniform distribution for generating standard deviations.}

\item{...}{Other arguments to \code{\link{rcor_mat}}.}
}
\value{
A covariance matrix with an attribute 'which_dims_cor', indicating 
which dimensions that are the correlated ones.
}
\description{
Generates a random correlation matrix by \eqn{R} \code{\link{rcor_mat}}, then
scales it to a covariance matrix by \eqn{CRC}, where
C is a diagonal matrix of uniformly generated standard deviations in
a specified range.
}
