% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpca_example_in_paper.R
\name{reproduce_tpca_example}
\alias{reproduce_tpca_example}
\title{Reproduces the figures in Section 2}
\usage{
reproduce_tpca_example()
}
\description{
This function does the following:
1) Runs the tpca-algorithm on the randomly drawn correlation 
matrix with 10^4 change simulations. 
(Call \code{get_example_cor_mat()} to get the correlation matrix.)
2) Creates ggplot objects for the sensitivity results and arranges them.
3) Saves two files to your current directory: ex_hellinger-d20.png and
ex_hellinger_prop-d20.png.
}
