% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_distributions.R
\name{set_uniform_cd}
\alias{set_uniform_cd}
\title{Title}
\usage{
set_uniform_cd(data_dim, prob = rep(1/3, 3), sparsities = 2:data_dim,
  mean_int = c(-1.5, 1.5), sd_int = c(2.5^(-1), 2.5),
  sd_inc_prob = 0.5, cor_int = c(0, 1))
}
\arguments{
\item{print_which}{}
}
\value{

}
\description{
Description
}
\details{
Detailed description.
}
\examples{

}
