% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpca.R
\name{tpca}
\alias{tpca}
\alias{tailorPCA}
\alias{tailoredPCA}
\title{Tailors the choice of principal axes to a set of distributional changes}
\usage{
tpca(cov_mat, change_distr = "full_uniform",
  divergence = "normal_hellinger", cutoff = 0.99,
  max_axes = ncol(cov_mat), n_sim = 10^3, print_which = TRUE)
}
\arguments{
\item{cov_mat}{A covariance matrix, i.e., a numeric matrix that is positive
definite.}

\item{change_distr}{A string or a change distribution object. A string can be
used to choose among a set of already implemented distributions:
'full_uniform', 'mean_only', 'sd_only', 'cor_only'. Custom change
distributions can be specified by using the \code{\link{set_uniform_cd()}}
function.}

\item{divergence}{A string specifying which divergence metric to use.}

\item{cutoff}{A numeric between 0 and 1 governing how many principal axes to
retain.}

\item{print_which}{}
}
\value{

}
\description{
Here goes the description.
}
\details{
Detailed description.
}
\examples{

}
