% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_distributions.R
\name{set_uniform_cd}
\alias{set_uniform_cd}
\title{Set a uniform change distribution.}
\usage{
set_uniform_cd(data_dim, prob = rep(1/3, 3), sparsities = 2:data_dim,
  mean_int = c(-1.5, 1.5), sd_int = c(2.5^(-1), 2.5),
  sd_inc_prob = 0.5, cor_int = c(0, 1), change_equal = FALSE)
}
\arguments{
\item{data_dim}{An integer specifying the dimension of the data.}

\item{prob}{A numeric vector of length 3 specifying the probability of a
change in the mean, variance or correlation, respectively.}

\item{sparsities}{An integer vector containing values between 2 (minimum 
sparsity for changes in the correlation) and data_dim.}

\item{mean_int}{A vector of length 2 specifying the lower and upper bound
of the interval that changes in the mean components are drawn uniformly from.}

\item{sd_int}{A vector of length 2 specifying the lower and upper bound
of the interval that multiplicative changes in the standard deviations
are drawn from. Must be between \eqn{\epsilon} and \eqn{\epsilon^{-1}}, where 
\eqn{\epsilon = } \code{sqrt(.Machine$double.eps)}.}

\item{sd_inc_prob}{A numeric between 0 and 1 indicating the probability of
a change in the standard deviation being an increase.}

\item{cor_int}{A vector of length 2 specifying the lower and upper bound
of the interval that multiplicative changes in the correlations are drawn 
from. Only values between 0 and 1 are allowed, i.e. a jump towards
0 of the correlation coefficients.}

\item{change_equal}{A logical for whether all affected dimensions should 
change by the same amount, or by amounts independent of eachother.
For example, if TRUE and the mean changes, \eqn{\mu_d = \mu} for
all \eqn{d} in \eqn{D}, where \eqn{D} is the set of affected dims.}
}
\value{
\code{set_uniform_cd} returns an S3 object of class "change_distr". 
This is a list containing the following functions:
\describe{
  \item{\code{draw_types(n_sim)}}{A function that draws \code{n_sim} change 
  types among ('mean', 'sd', 'cor') with probabilities \code{prob}.}
  \item{\code{draw_sparsities(n_sim)}}{A function that draws \code{n_sim} 
  change sparsities uniformly from \code{sparsities}.}
  \item{\code{draw_dims(k)}}{A function that uniformly draws a size k subset 
  of affected dimensions.}
  \item{\code{draw_mean(k)}}{A function that draws k values uniformly from
  \code{mean_int}.}
  \item{\code{draw_sd(k)}}{A function that draws an increase in the 
  standard deviation uniformly from [0, \code{sd_int[2]}] with probability 
  \code{sd_inc_prob}, and a decrease in standard deviation uniformly
  from [\code{sd_int[1]}, 1] with probability 1 - \code{sd_inc_prob}.}
  \item{\code{draw_cor(k)}}{A function that draws k values uniformly from
  \code{cor_int}}
}
The functions with argument \code{n_sim} are called with the number of 
simulations as arguments, while the ones with argument \code{k} are called
in each simulation run with a number depending on the change sparsity.
}
\description{
This function is used to specify a change distribution to be used in
conjuction with the function \code{\link{tpca}}. All components of the distribution
(marginal and conditional distributions) are uniform, but the
probability/importance of each type of change can be specified, along with
the range of sparsity of the change, as well as ranges for the sizes and
directions of the different change types. In each simulation run, after a
change sparsity has been drawn, which dimensions that are affected by a
change is always randomized. Choices for the distribution should reflect
prior knowledge about which changes that are of interest to detect.
}
\details{
See references.
}
\examples{

}
